+++ core/translate/emitter.rs
@@ -8,6 +8,7 @@ use limbo_sqlite3_parser::ast::{self, SortOrder};
 use tracing::{instrument, Level};
 
 use super::aggregation::emit_ungrouped_aggregation;
+use super::check_constraint::translate_check_constraint;
 use super::expr::{translate_condition_expr, translate_expr, ConditionMetadata};
 use super::group_by::{
     group_by_agg_phase, group_by_emit_row_phase, init_group_by, GroupByMetadata, GroupByRowSource,
@@ -1386,167 +1387,6 @@ fn emit_update_insns(
     }
 
     Ok(())
-}
-
-use crate::schema::Column;
-use limbo_sqlite3_parser::ast::Expr;
-fn translate_check_constraint(
-    program: &mut ProgramBuilder,
-    check_constraint_expr: &Expr,
-    registers_by_columns: &[(usize, &Column)],
-    jump_if_true: Option<BranchOffset>,
-    resolver: &Resolver,
-) -> usize {
-    match check_constraint_expr {
-        ast::Expr::Literal(_) => {
-            let reg = program.alloc_register();
-            let _ = translate_expr(program, None, &check_constraint_expr, reg, resolver);
-            return reg;
-        }
-        ast::Expr::Id(id) => {
-            for (reg, col) in registers_by_columns.iter() {
-                if col.name.as_ref().map_or(false, |name| *name == id.0) {
-                    return *reg;
-                }
-            }
-            unreachable!();
-        }
-        ast::Expr::Binary(lhs, ast::Operator::Equals, rhs) => {
-            use crate::vdbe::insn::CmpInsFlags;
-            let lhs_reg =
-                translate_check_constraint(program, lhs, registers_by_columns, None, resolver);
-            let rhs_reg =
-                translate_check_constraint(program, rhs, registers_by_columns, None, resolver);
-            if let Some(label) = jump_if_true {
-                program.emit_insn(Insn::Eq {
-                    lhs: lhs_reg,
-                    rhs: rhs_reg,
-                    target_pc: label,
-                    flags: CmpInsFlags::default().jump_if_null(),
-                    collation: program.curr_collation(),
-                });
-            }
-            return rhs_reg;
-        }
-        ast::Expr::Binary(lhs, ast::Operator::Add, rhs) => {
-            let lhs_reg =
-                translate_check_constraint(program, lhs, registers_by_columns, None, resolver);
-            let rhs_reg =
-                translate_check_constraint(program, rhs, registers_by_columns, None, resolver);
-            let dest_reg = program.alloc_register();
-            program.emit_insn(Insn::Add {
-                lhs: lhs_reg,
-                rhs: rhs_reg,
-                dest: dest_reg,
-            });
-            if let Some(label) = jump_if_true {
-                program.emit_insn(Insn::If {
-                    reg: dest_reg,
-                    target_pc: label,
-                    jump_if_null: true,
-                });
-            }
-            return dest_reg;
-        }
-        ast::Expr::Binary(lhs, ast::Operator::Subtract, rhs) => {
-            let lhs_reg =
-                translate_check_constraint(program, lhs, registers_by_columns, None, resolver);
-            let rhs_reg =
-                translate_check_constraint(program, rhs, registers_by_columns, None, resolver);
-            let dest_reg = program.alloc_register();
-            program.emit_insn(Insn::Subtract {
-                lhs: lhs_reg,
-                rhs: rhs_reg,
-                dest: dest_reg,
-            });
-            if let Some(label) = jump_if_true {
-                program.emit_insn(Insn::If {
-                    reg: dest_reg,
-                    target_pc: label,
-                    jump_if_null: true,
-                });
-            }
-            return dest_reg;
-        }
-        ast::Expr::Binary(lhs, ast::Operator::Multiply, rhs) => {
-            let lhs_reg =
-                translate_check_constraint(program, lhs, registers_by_columns, None, resolver);
-            let rhs_reg =
-                translate_check_constraint(program, rhs, registers_by_columns, None, resolver);
-            let dest_reg = program.alloc_register();
-            program.emit_insn(Insn::Multiply {
-                lhs: lhs_reg,
-                rhs: rhs_reg,
-                dest: dest_reg,
-            });
-            if let Some(label) = jump_if_true {
-                program.emit_insn(Insn::If {
-                    reg: dest_reg,
-                    target_pc: label,
-                    jump_if_null: true,
-                });
-            }
-            return dest_reg;
-        }
-        ast::Expr::Binary(lhs, ast::Operator::Divide, rhs) => {
-            let lhs_reg =
-                translate_check_constraint(program, lhs, registers_by_columns, None, resolver);
-            let rhs_reg =
-                translate_check_constraint(program, rhs, registers_by_columns, None, resolver);
-            let dest_reg = program.alloc_register();
-            program.emit_insn(Insn::Divide {
-                lhs: lhs_reg,
-                rhs: rhs_reg,
-                dest: dest_reg,
-            });
-            if let Some(label) = jump_if_true {
-                program.emit_insn(Insn::If {
-                    reg: dest_reg,
-                    target_pc: label,
-                    jump_if_null: true,
-                });
-            }
-            return dest_reg;
-        }
-        ast::Expr::Binary(lhs, ast::Operator::Modulus, rhs) => {
-            let lhs_reg =
-                translate_check_constraint(program, lhs, registers_by_columns, None, resolver);
-            let rhs_reg =
-                translate_check_constraint(program, rhs, registers_by_columns, None, resolver);
-            let dest_reg = program.alloc_register();
-            program.emit_insn(Insn::Remainder {
-                lhs: lhs_reg,
-                rhs: rhs_reg,
-                dest: dest_reg,
-            });
-            if let Some(label) = jump_if_true {
-                program.emit_insn(Insn::If {
-                    reg: dest_reg,
-                    target_pc: label,
-                    jump_if_null: true,
-                });
-            }
-            return dest_reg;
-        }
-        ast::Expr::Parenthesized(expr) => {
-            let dest_reg = translate_check_constraint(
-                program,
-                expr.get(0).expect("Expr should exist"),
-                registers_by_columns,
-                None,
-                resolver,
-            );
-            if let Some(label) = jump_if_true {
-                program.emit_insn(Insn::If {
-                    reg: dest_reg,
-                    target_pc: label,
-                    jump_if_null: true,
-                });
-            }
-            return dest_reg;
-        }
-        e => todo!("{}", &e),
-    }
 }
 
 /// Initialize the limit/offset counters and registers.
